latest version

+++ src/booking/service.py
@@ -1,21 +1,26 @@
 from datetime import date
+from typing import Optional
 from booking.repository import BookingRepository
 from booking.dto import BookingDto, BookingPostRequest
 from db.database import get_db
-from db.booking_status_enum import BookingStatusEnum
 from booking.mapper import db_booking_to_dto, booking_request_to_db
-from booking.exceptions import BookingNotFoundException, BookingStatusException, BookingException
+from booking.exceptions import BookingNotFoundException, BookingException
 from exceptions import InconsistentDatesException, DatesException
 from db.db_user import get_user
-from db.db_hotel import get_hotel
+from db.db_hotel import get_hotel_db
 from fastapi import HTTPException, status
-from db.models import DbUser, DbHotel
+from db.models import DbUser, DbHotel, DbRoom, DbBooking
+from schemas import UserBase
+from rooms.repository import RoomsRepository
+from sqlalchemy.orm import make_transient
 
 
 class BookingService:
-    def __init__(self):
+    def __init__(self, current_user: UserBase):
         self.db = next(get_db())
         self.repository = BookingRepository(self.db)
+        self.rooms_repository = RoomsRepository(self.db)
+        self.current_user = current_user
 
     def get_all(self):
         db_bookings = self.repository.get_all()
@@ -44,13 +49,28 @@ class BookingService:
             raise BookingNotFoundException(f"Booking with id {booking_id} not found.")
         return db_booking_to_dto(db_booking)
 
-    def add(self, booking: BookingPostRequest):
+    def add_current_user(self, booking: BookingPostRequest):
+        return self.add(booking, self.current_user)
+
+    def create_for_user(self, booking: BookingPostRequest, username: str):
+        db_current_user = get_user(self.db, self.current_user.username)
+        if not db_current_user.is_admin:
+            raise HTTPException(
+                status_code=status.HTTP_403_FORBIDDEN,
+                detail=f"User {self.current_user.username} is forbidden to book for another user"
+            )
+        user = get_user(self.db, username)
+        return self.add(booking, user)
+
+    def add(self, booking: BookingPostRequest, user: UserBase):
         check_dates(booking)
-        user = get_user(self.db, booking.client_username)
+        user = get_user(self.db, user.username)
         check_user_is_active(user)
         hotel = self.get_hotel_by_id(booking.hotel_id)
         check_hotel_is_active(hotel)
-        db_booking = booking_request_to_db(booking, user, hotel)
+        room = self.get_room(booking.room_id, hotel)
+        self.check_rooms_availability(booking)
+        db_booking = booking_request_to_db(booking, user, hotel, room)
         new_booking = self.repository.add(db_booking)
         self.db.commit()
         self.db.refresh(new_booking)
@@ -59,53 +79,36 @@ class BookingService:
 
     def update(self, id: int, booking: BookingPostRequest):
         current_booking = self.get_by_id(id)
-        check_status_is_draft(current_booking)
         check_dates(booking)
-        user = get_user(self.db, booking.client_username)
-        if user.username != current_booking.client_username:
-            raise BookingException(f"Cannot change username from "
-                                   f"'{current_booking.client_username}' to '{user.username}'")
+        current_db_booking = self.repository.get(id)
+        self.repository.remove(current_db_booking)
+        self.db.flush()
+        self.check_rooms_availability(booking)
+        current_db_booking = DbBooking()
+        current_db_booking.id = id
+        self.check_booking_authorization(id, self.current_user.username, 'update')
         hotel = self.get_hotel_by_id(booking.hotel_id)
         if hotel.id != current_booking.hotel.id:
             raise BookingException(f"Cannot change hotel from '{current_booking.hotel.name} "
                                    f"id={current_booking.hotel.id}' to '{hotel.name} id={hotel.id}'")
-        current_db_booking = self.repository.get(id)
-        booking_request_to_db(booking, user, hotel, current_db_booking)
+        room = self.get_room(booking.room_id, hotel)
+
+        user = get_user(self.db, self.current_user.username)
+        booking_request_to_db(booking, user, hotel, room, current_db_booking)
+        self.repository.add(current_db_booking)
         self.db.commit()
         self.db.refresh(current_db_booking)
         return db_booking_to_dto(current_db_booking)
 
     def remove(self, booking_id: int):
         booking = self.get_by_id(booking_id)
-        check_status_is_draft(booking)
+        self.check_booking_authorization(booking_id, booking.client_username, 'remove')
         db_booking = self.repository.get(booking_id)
         self.repository.remove(db_booking)
         self.db.commit()
 
-    def confirm(self, booking_id: int):
-        return self.set_next_status(booking_id, True)
-
-    def cancel(self, booking_id: int):
-        return self.set_next_status(booking_id, False)
-
-    def undo_confirm(self, booking_id: int):
-        booking = self.get_by_id(booking_id)
-        check_status_is_confirmed(booking)
-        return self.set_status(booking_id, BookingStatusEnum.draft)
-
-    def undo_cancel(self, booking_id: int):
-        booking = self.get_by_id(booking_id)
-        check_status_is_canceled(booking)
-        return self.set_status(booking_id, BookingStatusEnum.draft)
-
-    def set_next_status(self, booking_id: int, is_approve: bool):
-        booking = self.get_by_id(booking_id)
-        check_status_is_draft(booking)
-        status = BookingStatusEnum.confirmed if is_approve else BookingStatusEnum.canceled
-        return self.set_status(booking_id, status)
-
     def get_hotel_by_id(self, id: int):
-        hotel = get_hotel(self.db, id)
+        hotel = get_hotel_db(self.db, id)
         if hotel is None:
             raise HTTPException(
                 status_code=status.HTTP_404_NOT_FOUND,
@@ -113,38 +116,44 @@ class BookingService:
             )
         return hotel
 
-    def set_status(self, booking_id: int, status: BookingStatusEnum):
-        db_booking = self.repository.get(booking_id)
-        db_booking.status = status
-        self.db.commit()
-        self.db.refresh(db_booking)
-        return db_booking_to_dto(db_booking)
+    def check_rooms_availability(self, request: BookingPostRequest):
+        needed_rooms_amount = request.rooms_amount
+        available_count = self.rooms_repository.get_available_rooms_count(request)
+        if (available_count - needed_rooms_amount) < 0:
+            hotel = get_hotel_db(self.db, request.hotel_id)
+            raise HTTPException(
+                status_code=status.HTTP_400_BAD_REQUEST,
+                detail=f"Hotel '{hotel.name}' doesn't have enough rooms available "
+                       f"(needed count is {needed_rooms_amount} "
+                       f"and available count is {available_count})"
+            )
 
+    def check_booking_authorization(self, booking_id: int, booking_usernames: list[str], action: str):
+        if self.current_user.username not in booking_usernames:
+            raise HTTPException(
+                status_code=status.HTTP_403_FORBIDDEN,
+                detail=f"User {self.current_user.username} is forbidden to {action} booking with id={booking_id}"
+            )
 
-def check_status_is_draft(booking: BookingDto):
-    check_status(booking, BookingStatusEnum.draft)
-
-
-def check_status_is_confirmed(booking: BookingDto):
-    check_status(booking, BookingStatusEnum.confirmed)
-
-
-def check_status_is_canceled(booking: BookingDto):
-    check_status(booking, BookingStatusEnum.canceled)
-
-
-def check_status(booking: BookingDto, status: BookingStatusEnum):
-    if booking.status != status:
-        raise BookingStatusException(f"Booking should be {status.value}.")
+    def get_room(self, room_id: int, hotel: DbHotel):
+        room = self.rooms_repository.get(room_id)
+        if room is None:
+            raise BookingException(f"Room with id={room_id} not found")
+        if room.hotel_id != hotel.id:
+            raise BookingException(f"Room with id={room_id} does not belong to hotel {hotel.name} id={hotel.id}")
+        if not room.status:
+            raise BookingException(f"Room for {room.guests_count} guests in Hotel '{hotel.name}' is not available")
+        return room
 
 
 def check_dates(booking: BookingPostRequest):
-    if booking.date_start > booking.date_end:
-        raise InconsistentDatesException(f"Dates are inconsistent (start '{booking.date_start}' > end "
-                                         f"'{booking.date_end}')")
 
-    if booking.date_start < date.today():
-        raise DatesException(f"Impossible to make booking in the past ({booking.date_start} < {date.today()})")
+    if booking.checkin >= booking.checkout:
+        raise InconsistentDatesException(f"Dates are inconsistent (start '{booking.checkin}' >= end "
+                                         f"'{booking.checkout}')")
+
+    if booking.checkin < date.today():
+        raise DatesException(f"Impossible to make booking in the past ({booking.checkin} < {date.today()})")
 
 
 def check_user_is_active(user: DbUser):
